Issue: Cannot run github action in my github CI/CD due to missing node_modules directory.
Cause: yarn install or npm install isn't run prior to running gatsby cli on github CI/CD.
Fix: Ensuring to run yarn before running gatsby cli.
Note that we could do this manually as well, but it likely belongs here as having node_modules being installed is a dependency for this to work.
Test: Tested github action w/ the fix on my fork.